Reuploading sidebar-demo-files folder to display new CSS sheets and sidebar features
For the sidebar, when the user clicks an option, it scrolls to a specific area. Also, when the user scrolls, it highlights the appropriate option on the sidebar.

For CSS sheets, I updated the index.css, navbar.css, sidebar.css, and course.css to make the designs look more like Hertz's Figma prototype. I also added some nice fade-in transitions